+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [#508](https://github.com/teambit/bit/issues/508) structure do not support anything other than one dynamic param per folder
 - support yarn as package manager
 - add package manager config to bit.json
+- [#543](https://github.com/teambit/bit/issues/543) remove imported component not working
+
 
 ## [0.12.0-ext.11] - 2018-01-02